requirements.txt
APScheduler==3.10.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.9
django-apscheduler==0.6.2
django-extensions==3.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==5.2.2
idna==3.4
inflection==0.5.1
PyJWT==2.7.0
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
